Update Docker Image Metaflow Versions to current releases
Update Postgres to a more recent Versions that is not deprecated and therefore cheaper (No extended Support Fee).

Add Filesystem size variable and Launch Template for Batch, so it's configurable. (Had issues with running out of space after a certain number of flow steps. 30 GB is the current default, same as the aws default size)

Use current generation default values for Batch Env Instance Types.

Make 1 an allowed Value for Min and DesiredCPUBatch